fix(manyFolders): changing folder structure
