+++ src/controllers/requestController.js
@@ -10,28 +10,4 @@ async function createRequest(req, res) {
     }
 }
 
-async function getRequests(req, res) {
-    try {
-        const request = req.body;
-        const requests = await requestService.getRequests(request);
-        res.status(200).send({ requests });
-    } catch (err) {
-        res.status(500).send(err.message);
-    }
-}
-
-async function updateRequestStatus(req, res) {
-    try {
-        const { id, status, observations } = req.body;
-        const result = await requestService.updateRequestStatus({ id, status, observations });
-        res.status(200).send({ result });
-    } catch (err) {
-        res.status(500).send(err.message);
-    }
-}
-
-module.exports = {
-    createRequest,
-    getRequests,
-    updateRequestStatus
-};
+module.exports = { createRequest };
