+++ src/routes/adminRoutes.js
@@ -1,0 +1,7 @@
+const { router } = require('../config/server');
+const { getRequests } = require('../controllers/adminController');
+const adminRoutes = router;
+
+adminRoutes.get('/get', getRequests);
+
+module.exports = { adminRoutes };
