+++ src/services/adminService.js
@@ -1,0 +1,8 @@
+const adminModel = require("../models/admin");
+
+
+async function getRequests(request) {
+    return await adminModel.getRequests(request);
+};
+
+module.exports = { getRequests };
