+++ src/services/approverService.js
@@ -1,0 +1,7 @@
+const approverModel = require("../models/approver");
+
+async function updateRequestStatus(request) {
+    return await approverModel.updateRequestStatus(request);
+}
+
+module.exports = { updateRequestStatus };
